Remove _exception_handler.py
